# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Campus Platform API
  version: 1.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/errands:
    get:
      operationId: ListErrands
      tags: [errand]
      summary: "浏览待接跑腿任务"
      x-generated-module: errand
      x-permission: errand:read
      security: [{bearerAuth: []}]
      parameters:
        - name: page
          in: query
          required: false
          schema: {type: integer, format: int32}
        - name: page_size
          in: query
          required: false
          schema: {type: integer, format: int32}
      responses:
        "200":
          $ref: "#/components/responses/Success"
    post:
      operationId: CreateErrand
      tags: [errand]
      summary: "发布跑腿任务"
      x-generated-module: errand
      x-permission: errand:create
      security: [{bearerAuth: []}]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - reward_cents
                - pickup_location
                - dropoff_location
                - deadline
                - contact_type
                - contact
              properties:
                title:
                  type: string
                description:
                  type: string
                reward_cents:
                  type: integer
                  format: int64
                  minimum: 1
                pickup_location:
                  type: string
                dropoff_location:
                  type: string
                deadline:
                  type: string
                  format: date-time
                contact_type:
                  type: string
                contact:
                  type: string
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/errands/mine:
    get:
      operationId: ListMyErrands
      tags: [errand]
      summary: "查询我的跑腿任务"
      x-generated-module: errand
      x-permission: errand:read_mine
      security: [{bearerAuth: []}]
      parameters:
        - name: page
          in: query
          required: false
          schema: {type: integer, format: int32}
        - name: page_size
          in: query
          required: false
          schema: {type: integer, format: int32}
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /api/v1/errands/{id}:
    get:
      operationId: GetErrand
      tags: [errand]
      summary: "查询跑腿任务详情"
      x-generated-module: errand
      x-permission: errand:read_item
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: uint64, minimum: 1}}
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      operationId: UpdateErrand
      tags: [errand]
      summary: "修改未接单的跑腿任务"
      x-generated-module: errand
      x-permission: errand:update
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: uint64, minimum: 1}}
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - reward_cents
                - pickup_location
                - dropoff_location
                - deadline
                - expected_version
              properties:
                title:
                  type: string
                description:
                  type: string
                reward_cents:
                  type: integer
                  format: int64
                  minimum: 1
                pickup_location:
                  type: string
                dropoff_location:
                  type: string
                deadline:
                  type: string
                  format: date-time
                contact_type:
                  type: string
                contact:
                  type: string
                expected_version:
                  type: integer
                  format: uint64
                  minimum: 1
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/errands/{id}/accept:
    post:
      operationId: AcceptErrand
      tags: [errand]
      summary: "原子接单并创建交易订单"
      x-generated-module: errand
      x-permission: errand:accept
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: uint64, minimum: 1}}
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expected_version
              properties:
                expected_version:
                  type: integer
                  format: uint64
                  minimum: 1
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/errands/{id}/cancel:
    post:
      operationId: CancelErrand
      tags: [errand]
      summary: "取消跑腿任务"
      x-generated-module: errand
      x-permission: errand:cancel
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: uint64, minimum: 1}}
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expected_version
              properties:
                expected_version:
                  type: integer
                  format: uint64
                  minimum: 1
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/errands/{id}/complete:
    post:
      operationId: CompleteErrand
      tags: [errand]
      summary: "发布者确认跑腿完成"
      x-generated-module: errand
      x-permission: errand:complete
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: uint64, minimum: 1}}
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expected_version
              properties:
                expected_version:
                  type: integer
                  format: uint64
                  minimum: 1
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/Error"
...
  /api/v1/errands/{id}/pickup:
    post:
      operationId: PickupErrand
      tags: [errand]
      summary: "跑腿员确认取件"
      x-generated-module: errand
      x-permission: errand:pickup
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: uint64, minimum: 1}}
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expected_version
              properties:
                expected_version:
                  type: integer
                  format: uint64
                  minimum: 1
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
...
  responses:
    Success:
      description: Successful response
      content: {application/json: {schema: {$ref: "#/components/schemas/SuccessEnvelope"}}}
...
    Error:
      description: Error response
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [data, request_id]
      properties: {data: {}, request_id: {type: string}}
...
    ErrorEnvelope:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties: {code: {type: string}, message: {type: string}}
        request_id: {type: string}